requirements.txt
pyrogram==2.0.106
TgCrypto==1.2.5
pymongo==4.6.1
python-dotenv==1.0.0
requests==2.31.0
aria2p==0.11.3
aiohttp==3.9.1
motor==3.3.2
dnspython==2.4.2
shortzy==1.0.1